cryomech: Move cryomech cpa agent into library
