+++ docker-compose.yml
@@ -89,15 +89,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # Cryomech CPA compressor log agent
-  # --------------------------------------------------------------------------
-  ocs-cryomech-cpa-agent:
-    image: "ocs-cryomech-cpa-agent"
-    build: ./agents/cryomech_cpa/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # LabJack Agent
   # --------------------------------------------------------------------------
   ocs-labjack-agent:
